FRONT/Proyecto_ADVERT7/requirements.txt
django
pillow
django-admin-interface
django-crispy-forms
django-rest-framework
social-auth-app-django
django-pwa
django-bootstrap4
django-bootstrap-datepicker-plus
